Created Login, Logout, Signup possibilities

+++ backend/handlers.js
@@ -1,8 +1,10 @@
 const mongoose = require("mongoose");
 const House = require("./Schemas/HouseSchema");
 const { GridFSBucket, GridFSBucketReadStream, ObjectId } = require("mongodb");
-const crypto = require('crypto');
-require('dotenv').config();
+const crypto = require("crypto");
+const User = require("./schemas/UserSchema.js");
+const bcrypt = require("bcrypt");
+require("dotenv").config();
 var fs = require("fs");
 const AWS = require("aws-sdk");
 
@@ -14,9 +16,9 @@ AWS.config.update({
   accessKeyId: AWS_ACCESS_KEY_ID,
   secretAccessKey: AWS_SECRET_ACCESS_KEY,
 });
+
 const s3 = new AWS.S3();
 
-let gfs = null;
 async function connectToDatabase() {
   try {
     mongoose.connect(
@@ -29,11 +31,12 @@ async function connectToDatabase() {
     const conn = mongoose.connection;
     conn.once("open", () => {
       GridFSBucketReadStream.mongo = mongoose.mongo;
-      gfs = new GridFSBucket(conn.db, { bucketName: "uploads" });
+      new GridFSBucket(conn.db, { bucketName: "uploads" });
       console.log("Connected to MongoDB");
     });
   } catch (error) {
     console.error("Error connecting to MongoDB:", error);
+    throw error;
   }
 }
 
@@ -62,12 +65,13 @@ const uploadImages = async (req, res) => {
     };
 
     try {
-      const uploadResult = await s3.upload(params).promise();
-      fileIds.push(uploadResult.Key);
-
-      fs.unlink(files[fileIndex].path, () => {
-        uploadFile(fileIndex + 1);
-      });
+      let uploadResult;
+      if (null !== (uploadResult = await s3.upload(params).promise())) {
+        fileIds.push(uploadResult.Key);
+        fs.unlink(files[fileIndex].path, () => {
+          uploadFile(fileIndex + 1);
+        });
+      }
     } catch (error) {
       console.error("Error uploading file to S3:", error);
       res.status(500).json({ error: "Internal Server Error" });
@@ -112,13 +116,137 @@ const getAllHouses = async (req, res) => {
 };
 
 const generateRandomKey = () => {
-  const randomString = crypto.randomBytes(16).toString('hex');
+  const randomString = crypto.randomBytes(16).toString("hex");
   const timestamp = Date.now().toString();
   return `${randomString}-${timestamp}`;
 };
 
+const Login = async (req, res) => {
+  try {
+    const { email, password } = req.body;
+
+    const user = await User.findOne({ email });
+    if (!user) {
+      return res.status(401).json({ message: "Email not found" });
+    }
+
+    const isPasswordValid = await bcrypt.compare(password, user.password);
+
+    if (isPasswordValid) {
+      console.log("Password is correct");
+      return res.status(200).json({
+        message: "Login Successful",
+        user: {
+          _id: user._id,
+          fullName: user.fullName,
+          email: user.email,
+          photo: user.photo,
+          userCreatedDate: user.userCreatedDate,
+          houses: user.houses,
+        },
+      });
+    } else {
+      console.log("Password is incorrect");
+      return res.status(401).json({ message: "Incorrect Password" });
+    }
+  } catch (error) {
+    return res.status(500).json({ errors: "Internal Server Error" });
+  }
+};
+
+const Signup = async (req, res) => {
+  try {
+    const { fullName, email, password, photo } = req.body;
+
+    const existingUser = await User.findOne({ email });
+    if (existingUser) {
+      return res
+        .status(409)
+        .json({ message: "User with this email already exists" });
+    }
+
+    const saltRounds = 10;
+
+    // Hash the password
+    const hashedPassword = await bcrypt.hash(password, saltRounds);
+
+    const newUser = new User({
+      fullName,
+      email,
+      password: hashedPassword,
+      photo,
+      userCreatedDate: new Date(),
+      houses: [],
+    });
+
+    await newUser.save();
+
+    res.status(201).json({
+      message: "User created successfully",
+      user: {
+        _id: newUser._id,
+        fullName: newUser.fullName,
+        email: newUser.email,
+        photo: newUser.photo,
+        userCreatedDate: newUser.userCreatedDate,
+        houses: newUser.houses,
+      },
+    });
+  } catch (error) {
+    console.log(error);
+    res.status(500).json({ status: 500, message: "Internal Server Error" });
+  }
+};
+
+const GetUser = async (req, res) => {
+  try {
+    const userId = req.params.id; // Get the user ID from the URL parameter
+
+    const user = await User.findById(userId);
+
+    if (!user) {
+      return res.status(404).json({ error: "User not found" });
+    }
+
+    return res.status(200).json({
+      user: {
+        _id: user._id,
+        fullName: user.fullName,
+        email: user.email,
+        photo: user.photo,
+        userCreatedDate: user.userCreatedDate,
+        houses: user.houses,
+      },
+    });
+  } catch (error) {
+    console.error("An error occurred:", error);
+    return res.status(500).json({ error: "Internal Server Error" });
+  }
+};
+
+const DeleteUser = async (req, res) => {
+  try {
+    const { userId } = req.body; // Get the user ID from the request body
+
+    const deletedUser = await User.findByIdAndDelete(userId);
+
+    if (!deletedUser) {
+      return res.status(500).json({ message: "Failed to delete account" });
+    }
+
+    return res.status(200).json({ message: "Account deleted successfully" });
+  } catch (error) {
+    console.error("An error occurred:", error);
+    return res.status(500).json({ error: "Internal server error" });
+  }
+};
+
 module.exports = {
-  addHouse,
   uploadImages,
+  addHouse,
   getAllHouses,
+  Signup,
+  Login,
+  GetUser,
+  DeleteUser,
 };
